Enable synchronous access of app config (#4897)
