+++ scripts/appConfig.ts
@@ -1,103 +1,21 @@
-import {ISuperdeskGlobalConfig, IExtensions, IUser} from 'superdesk-api';
+import {ISuperdeskGlobalConfig, IExtensions} from 'superdesk-api';
 
-/* globals __SUPERDESK_CONFIG__: true */
-export const appConfig: ISuperdeskGlobalConfig = __SUPERDESK_CONFIG__;
+const appConfigForUnitTests = __SUPERDESK_CONFIG__;
+const userInterfaceLanguageForUnitTests = 'en';
 
-if (appConfig.startingDay == null) {
-    appConfig.startingDay = '0'; // sunday
-}
-
-if (appConfig.shortTimeFormat == null) {
-    appConfig.shortTimeFormat = 'HH:mm'; // 24h format
-}
-
-if (appConfig.ui == null) {
-    appConfig.ui = {};
-}
-
-if (appConfig.ui.sendEmbargo == null) {
-    appConfig.ui.sendEmbargo = false;
-}
-
-if (appConfig.ui.italicAbstract == null) {
-    appConfig.ui.italicAbstract = true;
-}
-
-if (appConfig.ui.publishEmbargo == null) {
-    appConfig.ui.publishEmbargo = true;
-}
-
-if (appConfig.authoring == null) {
-    appConfig.authoring = {};
-}
-
-if (appConfig.authoring.panels == null) {
-    appConfig.authoring.panels = {};
-}
-
-if (appConfig.authoring.panels.publish == null) {
-    appConfig.authoring.panels.publish = {};
-}
-
-if (appConfig.authoring.panels.publish.publishSchedule == null) {
-    appConfig.authoring.panels.publish.publishSchedule = true;
-}
-
-if (appConfig.authoring.panels.publish.publishingTarget == null) {
-    appConfig.authoring.panels.publish.publishingTarget = true;
-}
-
-if (appConfig.authoring.panels.sendTo == null) {
-    appConfig.authoring.panels.sendTo = {};
-}
-
-if (appConfig.authoring.panels.sendTo.publishSchedule == null) {
-    appConfig.authoring.panels.sendTo.publishSchedule = false;
-}
-
-const defaultDateFormat = 'MM/DD';
-const defaultTimeFormat = 'hh:mm';
-
-if (appConfig.view == null) {
-    appConfig.view = {
-        dateformat: defaultDateFormat,
-        timeformat: defaultTimeFormat,
-    };
-}
-
-if (appConfig.view.dateformat == null) {
-    appConfig.view.dateformat = defaultDateFormat;
-}
-
-if (appConfig.view.timeformat == null) {
-    appConfig.view.timeformat = defaultTimeFormat;
-}
-
-if (appConfig.longDateFormat == null) {
-    appConfig.longDateFormat = 'LLL';
-}
-
-if (appConfig.features == null) {
-    appConfig.features = {};
-}
-
-if (appConfig.features.autorefreshContent == null) {
-    appConfig.features.autorefreshContent = true; // default to true
-}
+export const appConfig: ISuperdeskGlobalConfig = window['appConfigLoaded'] ?? appConfigForUnitTests;
+export const userInterfaceLanguage: string = window['user-interface-language'] ?? userInterfaceLanguageForUnitTests;
 
 export const dashboardRoute = '/workspace';
 export const IDENTITY_KEY = 'sess:user';
 export const extensions: IExtensions = {};
 
+/**
+ * @deprecated
+ * only needed for compatibility with planning
+ */
 export function getUserInterfaceLanguage() {
-    const user: IUser | null = JSON.parse(localStorage.getItem(IDENTITY_KEY));
-    const language = user?.language ?? appConfig.default_language ?? window.navigator.language ?? 'en';
-
-    if (appConfig.profileLanguages?.includes(language)) {
-        return language;
-    } else {
-        return 'en';
-    }
+    return userInterfaceLanguage;
 }
 
 export const debugInfo = {
