+++ scripts/apps/users/directives/UserPreferencesDirective.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 /* tslint:disable:max-line-length */
 import {gettext} from 'core/utils';
-import {appConfig, extensions, getUserInterfaceLanguage} from 'appConfig';
+import {appConfig, extensions, userInterfaceLanguage} from 'appConfig';
 import {applyDefault} from 'core/helpers/typescript-helpers';
 import {DEFAULT_EDITOR_THEME} from 'apps/authoring/authoring/services/AuthoringThemesService';
 import {cloneDeep, pick} from 'lodash';
@@ -43,7 +43,7 @@ export function UserPreferencesDirective(
     return {
         templateUrl: asset.templateUrl('apps/users/views/user-preferences.html'),
         link: function(scope, element, attrs) {
-            const userLang = getUserInterfaceLanguage().replace('_', '-');
+            const userLang = userInterfaceLanguage.replace('_', '-');
             const body = angular.element('body');
             const NOTIFICATIONS_KEY = 'notifications';
