+++ scripts/core/decode-html-entities.ts
@@ -1,0 +1,7 @@
+export function decodeHtmlEntities(text: string): string {
+    const textArea = document.createElement('textarea');
+
+    textArea.innerHTML = text;
+
+    return textArea.value;
+}
