+++ scripts/core/get-superdesk-api-implementation.tsx
@@ -70,7 +70,7 @@ import {Icon} from './ui/components/Icon2';
 import {AuthoringWorkspaceService} from 'apps/authoring/authoring/services/AuthoringWorkspaceService';
 import ng from 'core/services/ng';
 import {Spacer, SpacerBlock, SpacerInlineFlex} from './ui/components/Spacer';
-import {appConfig, authoringReactViewEnabled, getUserInterfaceLanguage} from 'appConfig';
+import {appConfig, authoringReactViewEnabled, userInterfaceLanguage} from 'appConfig';
 import {httpRequestJsonLocal, httpRequestVoidLocal, httpRequestRawLocal} from './helpers/network';
 import {generatePatch} from './patch';
 import {getLinesCount} from 'apps/authoring/authoring/components/line-count';
@@ -521,7 +521,7 @@ export function getSuperdeskApiImplementation(
             },
             getRelativeOrAbsoluteDateTime: getRelativeOrAbsoluteDateTime,
             locale: {
-                code: getUserInterfaceLanguage().replace('_', '-'),
+                code: userInterfaceLanguage.replace('_', '-'),
                 firstDayOfWeek: appConfig.startingDay,
             },
         },
