+++ scripts/index.ts
@@ -34,20 +34,6 @@ import {dataStore} from 'data-store';
 
 let body = angular.element('body');
 
-function loadConfigs() {
-    return fetch(appConfig.server.url + '/client_config', {
-        method: 'GET',
-        mode: 'cors',
-    })
-        .then((res) => res.ok ? res.json() : Promise.reject())
-        .then((json) => {
-            merge(appConfig, json.config);
-
-            // allow e2e tests to overwrite appConfig via local storage
-            merge(appConfig, merge(appConfig, JSON.parse(localStorage.getItem('TEST_APP_CONFIG') ?? '{}')));
-        });
-}
-
 let started = false;
 
 function isDateFormatValid() {
@@ -173,46 +159,43 @@ export function startApp(
             },
         ]);
 
-    loadConfigs()
-        .then(() => {
-            if (isDateFormatValid() !== true) {
-                document.write('Invalid date format specified in config.view.dateFormat');
-                return;
+    if (isDateFormatValid() !== true) {
+        document.write('Invalid date format specified in config.view.dateFormat');
+        return;
+    }
+
+    /**
+     * @ngdoc module
+     * @name superdesk-client
+     * @packageName superdesk-client
+     * @description The root superdesk module.
+     */
+    angular.bootstrap(body, [
+        'superdesk.config',
+        'superdesk.core',
+        'superdesk.apps',
+        'superdesk.register_extensions',
+    ].concat(appConfig.apps || []), {strictDi: true});
+
+    window['superdeskIsReady'] = true;
+
+    body.attr('data-theme', 'dark-ui');
+
+    ng.waitForServicesToBeAvailable().then(() => {
+        if (sdApi.user.isLoggedIn()) {
+            if (appConfig.features.useTansaProofing) {
+                setupTansa();
             }
 
-            /**
-             * @ngdoc module
-             * @name superdesk-client
-             * @packageName superdesk-client
-             * @description The root superdesk module.
-             */
-            angular.bootstrap(body, [
-                'superdesk.config',
-                'superdesk.core',
-                'superdesk.apps',
-                'superdesk.register_extensions',
-            ].concat(appConfig.apps || []), {strictDi: true});
+            httpRequestJsonLocal<IRestApiResponse<ISubjectCode>>({method: 'GET', path: '/subjectcodes'})
+                .then(({_items}) => {
+                    store.dispatch({
+                        type: 'LOAD_SUBJECT_CODES',
+                        payload: keyBy(_items, ({qcode}) => qcode),
+                    });
+                });
 
-            window['superdeskIsReady'] = true;
-
-            body.attr('data-theme', 'dark-ui');
-
-            ng.waitForServicesToBeAvailable().then(() => {
-                if (sdApi.user.isLoggedIn()) {
-                    if (appConfig.features.useTansaProofing) {
-                        setupTansa();
-                    }
-
-                    httpRequestJsonLocal<IRestApiResponse<ISubjectCode>>({method: 'GET', path: '/subjectcodes'})
-                        .then(({_items}) => {
-                            store.dispatch({
-                                type: 'LOAD_SUBJECT_CODES',
-                                payload: keyBy(_items, ({qcode}) => qcode),
-                            });
-                        });
-
-                    registerGlobalKeybindings();
-                }
-            });
-        });
+            registerGlobalKeybindings();
+        }
+    });
 }
